moved listApis to shared_hooks.ts
