+++ projects/ui/src/Apis/gg_hooks.ts
@@ -16,18 +16,11 @@ import {
   User,
   isSubscriptionsListError,
 } from "./api-types";
-import { useSwrWithAuthListApis } from "./list-apis-helpers";
 import { fetchJSON, useMultiSwrWithAuth, useSwrWithAuth } from "./utility";
 
 //
 // Queries
 //
-
-export function useListApis() {
-  // Listing API's is more complicated since we support GMG and GG Portal Server APIs.
-  // This uses a custom SWR fetcher function which may do several fetch calls.
-  return useSwrWithAuthListApis();
-}
 
 // User
 export function useGetCurrentUser() {
