+++ projects/ui/src/Apis/shared_hooks.ts
@@ -1,0 +1,11 @@
+import { useSwrWithAuthListApis } from "./list-apis-helpers";
+
+//
+// Queries
+//
+
+export function useListApis() {
+  // Listing API's is more complicated since we support GMG and GG Portal Server APIs.
+  // This uses a custom SWR fetcher function which may do several fetch calls.
+  return useSwrWithAuthListApis();
+}
