+++ projects/ui/src/Components/Apis/gloo-gateway-components/ApisTab/ApisList.tsx
@@ -1,5 +1,5 @@
 import { useContext, useMemo } from "react";
-import { useListApiProducts } from "../../../../Apis/hooks";
+import { useListApiProducts } from "../../../../Apis/gg_hooks";
 import { AppContext } from "../../../../Context/AppContext";
 import { FilterPair, FilterType } from "../../../../Utility/filter-utility";
 import { EmptyData } from "../../../Common/EmptyData";
