+++ projects/ui/src/Components/Apps/Modals/CreateNewAppModal.tsx
@@ -2,7 +2,7 @@ import { CloseButton, Flex } from "@mantine/core";
 import { FormEvent, useEffect, useRef, useState } from "react";
 import toast from "react-hot-toast";
 import { di } from "react-magnetic-di";
-import { useCreateAppMutation, useListTeams } from "../../../Apis/hooks";
+import { useCreateAppMutation, useListTeams } from "../../../Apis/gg_hooks";
 import { FormModalStyles } from "../../../Styles/shared/FormModalStyles";
 import { Button } from "../../Common/Button";
 import { Loading } from "../../Common/Loading";
