+++ projects/ui/src/Components/Common/SubscriptionsList/SubscriptionInfoCard/Modals/ApproveSubscriptionModal.tsx
@@ -3,7 +3,7 @@ import { FormEvent } from "react";
 import toast from "react-hot-toast";
 import { di } from "react-magnetic-di";
 import { Subscription } from "../../../../../Apis/api-types";
-import { useAdminApproveSubscriptionMutation } from "../../../../../Apis/hooks";
+import { useAdminApproveSubscriptionMutation } from "../../../../../Apis/gg_hooks";
 import { FormModalStyles } from "../../../../../Styles/shared/FormModalStyles";
 import { Button } from "../../../Button";
