+++ projects/ui/src/Components/Structure/HeaderSectionLoggedIn.tsx
@@ -4,7 +4,7 @@ import { Popover } from "@mantine/core";
 import { useContext, useEffect, useMemo, useState } from "react";
 import { di } from "react-magnetic-di";
 import { NavLink, useLocation, useSearchParams } from "react-router-dom";
-import { useGetCurrentUser } from "../../Apis/hooks";
+import { useGetCurrentUser } from "../../Apis/gg_hooks";
 import { Icon } from "../../Assets/Icons";
 import { AppContext } from "../../Context/AppContext";
 import { AuthContext } from "../../Context/AuthContext";
