+++ projects/ui/src/Components/Teams/Details/Modals/DeleteTeamModal.tsx
@@ -4,7 +4,7 @@ import toast from "react-hot-toast";
 import { di } from "react-magnetic-di";
 import { useNavigate } from "react-router-dom";
 import { Team } from "../../../../Apis/api-types";
-import { useDeleteTeamMutation } from "../../../../Apis/hooks";
+import { useDeleteTeamMutation } from "../../../../Apis/gg_hooks";
 import { FormModalStyles } from "../../../../Styles/shared/FormModalStyles";
 import { Button } from "../../../Common/Button";
